refactor: arreglar nombre de modales de info de usuarios y clientes
